requirements.txt
numpy>=1.14.0
opencv-python>=3.4.0
tensorflow-gpu>=2.0.0-alpha0